+++ lib/src/style/dot_tab_indicator.dart
@@ -41,7 +41,9 @@ class _DotTabIndicatorPainter extends BoxPainter {
   ) : super(onChanged);
   @override
   void paint(Canvas canvas, Offset offset, ImageConfiguration configuration) {
-    final double yOffset = tabAlignment == TabAlignment.bottom ? configuration.size.height - radius * 2 : radius * 2;
+    final double yOffset = tabAlignment == TabAlignment.bottom
+        ? configuration.size.height - radius * 2
+        : radius * 2;
 
     final Offset circleOffset = offset +
         Offset(