+++ lib/src/widgets/sura_async_button.dart
@@ -127,7 +127,9 @@ class _SuraAsyncButtonState extends State<SuraAsyncButton> {
       ],
     );
 
-    final Widget loadingWidget = widget.loadingWidget ?? SuraTheme.of(context)?.buttonLoadingWidget ?? _buildLoadingWidget();
+    final Widget loadingWidget = widget.loadingWidget ??
+        SuraTheme.of(context)?.buttonLoadingWidget ??
+        _buildLoadingWidget();
 
     return Container(
       height: widget.height,
@@ -143,7 +145,9 @@ class _SuraAsyncButtonState extends State<SuraAsyncButton> {
                 : null,
         child: ConditionalWidget(
           condition: isLoading,
-          onTrue: () => widget.loadingType == LoadingType.Disable ? buttonContent : loadingWidget,
+          onTrue: () => widget.loadingType == LoadingType.Disable
+              ? buttonContent
+              : loadingWidget,
           onFalse: () => buttonContent,
         ),
         style: ElevatedButton.styleFrom(